rubocop: module/method too long
